created data folder and moved sellerModle to types folder
