+++ src/components/pages/sellers/sellersListDisplay/SellersListDisplay.tsx
@@ -1,0 +1,20 @@
+import type { JSX } from "react";
+import type { Seller } from "../../../../types/sellerModel";
+
+function SellersListDisplay({ sellers }: { sellers: Seller[] }) {
+
+    const sellerListItems: JSX.Element[] = sellers.map((seller) => {
+        return (
+            <SellerCard 
+                seller={seller}
+                key={seller.id}
+            />
+        );
+    });
+
+    return (
+        <div className="sellers-list">
+            {sellerListItems}
+        </div>
+    );
+}
