+++ src/types/sellerModel.ts
